requirements.txt
astroid==2.5.1
colorama==0.4.4
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysql-connector-python==8.0.23
pylint==2.7.2
toml==0.10.2
wrapt==1.12.1
